컨트롤러 설명 추가

+++ src/main/java/codehanzoom/greenwalk/user/domain/User.java
@@ -48,7 +48,7 @@ public class User {
 
     private int totalTrashCount;
 
-    private float totalWalkingDistance;
+    private double totalWalkingDistance;
 
     @JsonIgnore
     @OneToMany(mappedBy = "user")
